==> 03-cleaning.Rmd <==
# Missing values
First, we load the data and show the first couple lines here.
```{r}
restaurant_df <- read.csv(file = 'opentable.csv', encoding = "UTF-8")

head(restaurant_df,3)
```
Then, we check the number of NA's in each column. 

```{r}
restaurant_df[restaurant_df==""] <- NA
colSums(is.na(restaurant_df))
```


There are 3 NA's in website column. To find out where they are, we plot the missing patterns in data by rows with `mi` package.

```{r, echo=FALSE,warning=FALSE,message=FALSE,error=FALSE, results='hide',fig.keep='all'}
library(mi)

x <- invisible(missing_data.frame(restaurant_df))

image(x)

```

From the plot above, we are able to detect all the missing values are. Since it only consist of a small part of our data, we decided to remove these rows by using `na.omit()` function. And we check again to make sure there are no NA's in the dataset. 

```{r}
restaurant_df <- na.omit(restaurant_df)
colSums(is.na(restaurant_df))
```



```{r}
write.csv(restaurant_df, './restaurant.csv', row.names=FALSE)
```

We then check for NA's in our second dataset.

```{r}
restaurant_grades = read.csv(file = 'Restaurant_Grades.csv', encoding = "UTF-8")

head(restaurant_grades)
```
Here are the missing values.

```{r}
restaurant_grades[restaurant_grades==""] <- NA
colSums(is.na(restaurant_grades))
```

Since our dataset is large enough, we can delete these rows. Then double check.

```{r}
restaurant_grades <- na.omit(restaurant_grades)
colSums(is.na(restaurant_grades))
```       

```{r}
write.csv(restaurant_grades, './violations.csv', row.names=FALSE)
```
