==> 04-missing.Rmd <==
# Data Transformation

We'd like to order the data by `average_review` to see the general ranking of the restaurants.

```{r}
data <- read.csv("restaurant.csv", encoding = "UTF-8")
data <- data[order(data$average_review, decreasing = TRUE),]
head(data)
```
We'd like to join our datasets to see both the violations and the ratings. To accomplish that, we first convert all restaurants names to lower cases with `tolower()`, then strip symbols off on two datasets with `sub()` to create uniform syntax for names. After that, we used `sqldf` package to count the number of violations each restaurant has and join the two datasets together as our final dataset.

```{r}

data$name <- tolower(data$name)
data$name<-sub("\\-.*", "", data$name)

violations <- read.csv(file = 'violations.csv')

violations$DBA <- tolower(violations$DBA)
violations$DBA <- sub("\\(.*",'',violations$DBA)

library(sqldf)

violation_counts <- sqldf('select DBA,count(DBA) as violation_count from violations group by DBA')

df <- sqldf('select * from data as d join violation_counts as v on d.name = v.DBA')
```
We can take a glance at the final dataset.

```{r}
df$DBA <- NULL

head(df,3)
write.csv(df, './finaldf.csv', row.names=FALSE)
```

While the above final dataset contains information about restaurants which had violation records, in order to faciliate the interaction part of this project, we are going to create `cleaned_restaurants` and `cleaned_violations` datasets, so that we can 1.find each restaruant's violation total count (including 0) 2.boil down the large original violations file so the new file only contains restaurant week's restaurants' specific violation information.

```{r}
cleaned_restaurants_df <-
  sqldf('SELECT d.*,
  CASE WHEN v.violation_count IS NULL THEN 0 ELSE v.violation_count END AS violation_count
  FROM data AS d 
  LEFT JOIN violation_counts AS v 
  ON d.name = v.DBA')

cleaned_violations_df <-
  sqldf('SELECT *
  FROM violations v
  WHERE v.DBA IN (SELECT d.name FROM data AS d)')

write.csv(cleaned_restaurants_df, './docs/cleaned_restaurants.csv', row.names=FALSE)
write.csv(cleaned_violations_df, './docs/cleaned_violations.csv', row.names=FALSE)
```
