==> 05-results.Rmd <==
# Results

```{r}
df <- read.csv('finaldf.csv',encoding = "UTF-8")
ot <- read.csv('opentable.csv', encoding = "UTF-8")
```

## What are the restaurant types?

First, we'd like to see what are the main types of restaurants in NYC Restaurant Week. The following is a plot of the number count for each restaurant type.

```{r}
library(ggplot2)
library(tidyverse)

ggplot(ot, aes(x = fct_relevel(fct_rev(fct_infreq(restaurant_main_type)),'Other'))) + 
  geom_histogram(stat = 'count', binwidth=1, fill='lightblue') +
  theme_bw() +
  coord_flip()+
  ggtitle("Number of each restaurant type") +
  labs(x = 'restaurant type')
  
```

Not to our surprise, Italian restaurants are the most common restaurants in NYC Restaurant Week, followed by American and Steakhouse. 

How expensive are these restaurants? To explore this question, we plotted the following mosaic plot.

```{r,fig.height=6}
library(vcd)
library(vcdExtra)

mosaic(price_range ~ restaurant_main_type, data = df,rot_labels=c(0,90,0,0) ,
       highlighting_fill = c("sandybrown","lightblue","saddlebrown"),
       main = 'Price range of each restaurant type',
       gp_varnames = gpar(fontsize = 14, fontface = 1),
       gp_labels = gpar(fontsize = 10),
       labeling_args = list(offset_varnames = c(left = 2, top=0))) 

```

We can see that most of the restaurants attending restaurant week cost \$30 to \$50 per person. Surprisingly, French restaurants are not more expensive than other types of restaurants. In fact, there are bit any French restaurants costing more than $50 per person, whereas American restaurants have the most fancy restaurants.

## How are average rating and inspection violation counts related?

Then, we try to explore the relationship between restaurant average rating and inspection violation counts. To do that, we used `violation` created in data transformation, which contains both the restaurant ratings and the number of violations to plot the following scatterplot. Each restaurant is colored by their type.

```{r, fig.width=10}
ggplot(df, aes(x=average_review,y=violation_count, color = restaurant_main_type)) +
  geom_point() +
  ggtitle("Average rating vs. Violation count")+
  labs(x='average_rating') +
  geom_label(aes(4.3,34,label='obao'),color='red',size = 5, alpha = 0.5) +
  geom_label(aes(3.88,6,label='brasserie seoul'),color='red',size = 5, alpha = 0.5)

```

There is a major cluster in the bottom-right section of the graph, from average review of 4.4 to 4.75, and violation count from 4 to 8. This cluster represents that most of the restaurants attending restaurant week have relatively high ratings and few violations. However, the graph does not show any linear relationship between average rating and violations. It is also noticeable that there is an outlying restaurant with many violations but relatively low rating, so we think customers should pay special attention to that restaurant. We also labeled the worst rated restaurant in our data.

## Which restaurants have the most violations?

Meanwhile, we are interested in what types of restaurants have the most violations. So we plotted the following Cleveland dot plot.

```{r}
rest_type <- aggregate(df$violation_count, by= list(df$restaurant_main_type), FUN =mean)
rest_count <- aggregate(df$restaurant_type, by= list(df$restaurant_main_type), FUN = length)
rest_type <- merge(rest_type, rest_count, by = 'Group.1',all = TRUE)

colors <- c('avg'='red','count'='blue')

rest_type = rest_type %>%
  mutate(Type = fct_relevel(fct_reorder(Group.1, x.x),'Other')) 
  ggplot(rest_type,(aes(y=Type))) +
  geom_point(aes(x=x.x),color='red') +
  ggtitle('Average violation counts by type') +
  labs(x='average violation',y='restaurant type')
  #geom_point(aes(x=x.y,color='count')) +
  #labs(x=color = 'Legend') +
  #scale_color_manual(values = colors)



```

As seen from above Southeast Asian restaurants have the most average violations over all the restaurant. We believe it is a reasonable result given the size of our data set and that Obao, the Southeast Asian restaurant, has abnormally more violation than other restaurants. Other than that, Seafood and Mexican restaurants have the second and third most violations. Most types of restaurants have around 5 to 10 violations. 

By the way, we are giving  a big shoutout to Italian restaurants. They have the most numbers of restaurants, but lower average violations to most types of restaurants. Good job keeping that dishes clean, Italian restaurants!

## Where are these restaurants?

After looking at the cuisine pattern of the violations, we'd like to see if there are any geographical patters. That is, which neighborhoods have the most violations.

```{r}
library(dplyr)

Manhhattan_zips <- c(10026, 10027, 10030, 10037, 10039, 10001, 10011, 10018, 10019, 10020, 10036,	10029, 10035,10010, 10016, 10017, 10022,
                     10012, 10013, 10014,	10004, 10005, 10006, 10007, 10038, 10280,	10002, 10003, 10009, 10021, 10028, 10044, 10065, 10075, 10128,
                     10023, 10024, 10025, 10031, 10032, 10033, 10034, 10040)

df_man <- df[df$postal_code %in% Manhhattan_zips,] 

df_man <- df_man %>%
  mutate(Neighborhood = case_when(
    postal_code %in% c(	10026, 10027, 10030, 10037, 10039) ~ "Central Harlem",
    postal_code %in% c(10001, 10011, 10018, 10019, 10020, 10036) ~ 'Chelsea and Clinton',
    postal_code %in% c(10029, 10035) ~ 'East Harlem',
    postal_code %in% c(10010, 10016, 10017, 10022) ~ 'Gramercy Park and Murray Hill',
    postal_code %in% c(10012, 10013, 10014) ~ 'Greenwich Village and Soho',
    postal_code %in% c(10004, 10005, 10006, 10007, 10038, 10280) ~ 'Lower Manhattan',
    postal_code %in% c(10002, 10003, 10009) ~ 'Lower East Side ',
    postal_code %in% c(10021, 10028, 10044, 10065, 10075, 10128) ~ 'Upper East Side',
    postal_code %in% c(10023, 10024, 10025) ~ 'Upper West Side',
    postal_code %in% c(10031, 10032, 10033, 10034, 10040) ~ 'Inwood and Washington Heights'
  ))

avg_violation <- aggregate(df_man$violation_count, by = list(df_man$Neighborhood), FUN = mean)

ggplot(avg_violation,aes(fct_rev(reorder(Group.1,-x,sum)),x))+
  geom_col(binwidth=1, fill='lightblue') +
  theme_bw() +
  coord_flip() +
  ggtitle("Average vioaltion count by neighborhood") +
  labs(x = 'Neighborhood',y = 'average violation count')

```

From the plot above, it is clear that Central Harlem has the most average violations among all neighborhoods, followed by Greenwich Village and Soho. Upper East Side, with no doubt, has the cleanest and safest restaurant.

If we combine restaurant types into the each neighborhood, we get the following graph.

```{r,fig.height=6}
library(colorRamps)

df_man %>%
  group_by(Neighborhood, restaurant_main_type) %>%
  summarize(count = n()) %>%
  ggplot(aes(fct_rev(fct_infreq(Neighborhood)),count)) +
  geom_bar(aes(fill = restaurant_main_type), position = "stack", stat = "identity",alpha = 0.8) +
  scale_fill_manual(values =  primary.colors(21)) +
  coord_flip() +
  ggtitle('Total restaurants in each neighborhood by type') +
  labs(x='Neighborhood')
```

From the new graph, we are able to get a lot of new information. For example, Central Harlem has the most average violation count because there is only one restaurant there and it had a lot of violations. We are also able to see the restaurant diversity in each neighborhood. Chelsea and Clinton, for instance, has the most total violations, but it also consists of the most restaurant types, almost every type we find. So if you are looking for diversity of tastes, Chelsea and Clinton can be a good neighborhood to go.
