==> 07-conclusion.Rmd <==
# Conclusion

Our project is able to provide detailed individual restaurants information for users.While the interactive plotting page provides violation and review information of restaurants, some higher level summaries are available in our result page. When we are constructing these pages, we did encounter some problems that we were not able to solve at the current stage. 

First, our d3 SVG could not automatically scale when there is a change in the window size. Since SVG is not responsive to the size change, we have to manually adjust the size many times to fit the bookdown view. The second problem we had was the coloring problem in our result page. Since there are so many restaurant types, the current color palette does not have enough colors for us. We had to define a function to randomly select colors for restaurant types which made us lose the control over colors we want to use. Our potential improvement will focus on these two problems. We want to make the SVG responsive to the window size change. We will do some investigations on the color palette to gain more control over the colors. Another problem is that when we were plotting the last stacked bar chart, there were two bars couldn't be ordered by their length.

There are a lot of future direction for our project. For example, we could design an App that tracks restaurant inspection violations in real time, and notify the user if the restaurant they plan to dine in has a lot of violations. Data-wise, we could explore a larger dataset consisting most major cities on the East Coast to provide a more comprehensive analysis and visualization.
